requirements.txt
requests
beautifulsoup4

qdrant-client

sentence-transformers
torch

scikit-learn
numpy

fastapi
uvicorn
python-dotenv
pyjwt
passlib
bcrypt==3.2.2
flask
flask-cors
python-dotenv
google-genai
